Updated my init.el to current up-to-date version
Removed the usage of a mixture of Eglot and LSP to purely use LSP.

Now it only supports Python (JEDI) and C/C++ (CLANG), and BASH through bash-language-server package. If LSP detects it in your system it'll configure itself to work with it.

Migrated to using use-package for most if not all packages instead of required.

Organized the init.el to look better and shortened it more too.

Various others misc. stuff that I might've forgotten...